added QueryComponents interface to abstract out GetComponents
